+++ floodmodeller_api/test/test_logs_lf.py
@@ -1,9 +1,12 @@
 from pathlib import Path
+from unittest.mock import MagicMock, patch
 
 import pandas as pd
 import pytest
+from freezegun import freeze_time
 
 from floodmodeller_api import IEF, LF1
+from floodmodeller_api.logs import init_log_file
 
 
 @pytest.fixture
@@ -46,3 +49,54 @@ def test_lf1_from_ief(lf1_fp, test_workspace):
     assert lf1._filepath == lf1_from_ief._filepath
     assert lf1.info == lf1_from_ief.info
     pd.testing.assert_frame_equal(lf1.to_dataframe(), lf1_from_ief.to_dataframe())
+
+
+def test_log_file_unsupported(capsys):
+    lf = init_log_file(None, "lf1", True)
+
+    assert lf is None
+    assert (
+        capsys.readouterr().out
+        == "No progress bar as only 1D unsteady runs are supported. Simulation will continue as usual.\n"
+    )
+
+
+@pytest.mark.usefixtures("log_timeout")
+def test_log_file_timeout(capsys):
+    lf_filepath = MagicMock()
+    lf_filepath.is_file.return_value = False
+    lf = init_log_file(lf_filepath, "lf1", False)
+
+    assert lf is None
+    assert (
+        capsys.readouterr().out
+        == "No progress bar as log file is expected but not detected. Simulation will continue as usual.\n"
+    )
+
+
+@pytest.mark.usefixtures("log_timeout")
+@freeze_time("1970-01-01 00:00:00", tick=True)
+def test_log_file_from_old_run(capsys):
+    lf_filepath = MagicMock()
+    lf_filepath.is_file.return_value = True
+    lf_filepath.stat.return_value.st_mtime = -10
+    lf = init_log_file(lf_filepath, "lf1", False)
+
+    assert lf is None
+    assert (
+        capsys.readouterr().out
+        == "No progress bar as log file is from previous run. Simulation will continue as usual.\n"
+    )
+
+
+@pytest.mark.usefixtures("log_timeout")
+@freeze_time("1970-01-01 00:00:00", tick=True)
+def test_log_file_found():
+    lf_filepath = MagicMock()
+    lf_filepath.is_file.return_value = True
+    lf_filepath.stat.return_value.st_mtime = -1
+    with patch("floodmodeller_api.logs.lf.LF1") as lf1:
+        lf = init_log_file(lf_filepath, "lf1", False)
+
+    assert lf is not None
+    lf1.assert_called_once_with(lf_filepath, False)
